car-detection/requirements.txt
ultralytics==8.0.200
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.0.1
matplotlib==3.8.1
pyyaml==6.0.1
